feat: 添加多种几何体、SSAA和渲染计时功能
1. 新增Disk、Triangle求交类与单元测试，重构Intersection结构体
2. 重构相机射线生成接口，支持浮点像素坐标
3. 添加超采样抗锯齿(SSAA)功能，支持自定义采样数
4. 新增渲染时长格式化与窗口标题自动更新功能
5. 完善CMake与VS项目配置，新增对应测试项
6. 调整默认渲染分辨率与相机参数

+++ source/Camera.cpp
@@ -92,15 +92,15 @@ const glm::mat4& Camera::GetInverseCombinedMatrix() const
     return m_inverseCombinedMatrix;
 }
 
-Ray Camera::GetRay(int x, int y) const
+Ray Camera::GetRay(float x, float y) const
 {
     Ray ray;
     ray.origin = m_position;
 
     // 屏幕坐标的 Z 为 0，表示位于近裁剪面。
     const glm::vec4 screenPosition(
-        static_cast<float>(x),
-        static_cast<float>(y),
+        x,
+        y,
         0.0f,
         1.0f);
     glm::vec4 worldPosition = m_inverseCombinedMatrix * screenPosition;
